added data loading from csv, and header

+++ project_covid19.py
@@ -1,3 +1,12 @@
+# ----------------------------------------------------------------------
+# DATE: 2020/08/10
+# AUTHOR: Jerry Liu
+# EMAIL: Liuyal@sfu.ca
+#
+# DESCRIPTION:
+# COVID-19 Data Mining Project
+# ----------------------------------------------------------------------
+
 import os
 import sys
 import time
@@ -19,11 +28,20 @@ def delete_folder(path):
     shutil.rmtree(path, ignore_errors=True)
 
 
-def check_data(repos):
+def load_csv_data(file_name):
+    data_output = []
+    file = open(file_name, "r+", encoding="utf-8")
+    data = csv.reader(file)
+    for row in data: data_output.append(row)
+    file.close()
+    return data_output
+
+
+def check_repo_data(repos):
     if not os.path.exists(os.getcwd() + os.sep + "data"):
 
-        for folder in os.listdir(os.getcwd() ):
-            if os.path.isdir(folder) and ".idea" not in folder and ".git" not in folder:
+        for folder in os.listdir(os.getcwd()):
+            if os.path.isdir(folder) and ".idea" not in folder and ".git" not in folder and "data" not in folder:
                 delete_folder(os.getcwd() + os.sep + folder)
 
         for repo in repos: os.system("git clone " + repo)
@@ -33,7 +51,7 @@ def check_data(repos):
             if "twitter" in repo:
                 src = os.getcwd() + os.sep + repo.split('/')[-1].split('.')[0] + os.sep + "dailies"
                 dst = os.getcwd() + os.sep + "data" + os.sep + repo.split('/')[-1].split('.')[0]
-                shutil.copytree(src, dst, ignore=shutil.ignore_patterns("*.tsv.gz"))
+                shutil.copytree(src, dst, ignore=shutil.ignore_patterns("*.tsv.gz", "*.md"))
             else:
                 src = os.getcwd() + os.sep + repo.split('/')[-1].split('.')[0] + os.sep + "latest_data"
                 dst = os.getcwd() + os.sep + "data" + os.sep + repo.split('/')[-1].split('.')[0]
@@ -46,8 +64,38 @@ def check_data(repos):
             delete_folder(os.getcwd() + os.sep + repo.split('/')[-1].split('.')[0])
 
 
+def load_data(path):
+    location_data = []
+    twitter_data = {}
+    for folder in os.listdir(path):
+        if "twitter" in folder:
+            for sub_folder in os.listdir(path + os.sep + folder):
+                for files in os.listdir(path + os.sep + folder + os.sep + sub_folder):
+                    type = files.split("top1000")[-1].split('.')[0].replace("ii","i")
+                    if type not in twitter_data.keys():
+                        twitter_data[type] = {}
+                    else:
+                        if sub_folder not in twitter_data[type].keys():
+                            twitter_data[type][sub_folder] = []
+                    twitter_daily_data = load_csv_data(path + os.sep + folder + os.sep + sub_folder + os.sep + files)
+                    twitter_data[type][sub_folder] = twitter_daily_data
+        else:
+            for file in os.listdir(path + os.sep + folder):
+                if ".csv" in file:
+                    location_data = load_csv_data(path + os.sep + folder + os.sep + file)
+
+    return twitter_data, location_data
+
+
 if __name__ == "__main__":
     covid19_twitter_repo = r"https://github.com/thepanacealab/covid19_twitter.git"
     nCoV2019_location_repo = r"https://github.com/beoutbreakprepared/nCoV2019.git"
 
-    check_data([covid19_twitter_repo, nCoV2019_location_repo])
+    print("Checking COVID-19 GIT REPO data...")
+    check_repo_data([covid19_twitter_repo, nCoV2019_location_repo])
+
+    print("Loading COVID-19 Twitter and Location data...")
+    twitter_data, location_data = load_data(os.getcwd() + os.sep + "data")
+
+
+    print("EOS")
